Fix precision loss issues
The naive application of the inverse Möbius transformation was causing a
massive loss of precision. I believe it was numerically ill conditioned,
although there may be some SymPy bugs as well.

Instead, we have to apply the transformation using special Poly methods
(particularly, shift, which does a Taylor shift). See the comments in the code
for details.

The test case of n = 17 (which is the only one I've OCRed so far) matches the
20 digits from the paper exactly (except for some different rounding in the
last digits). I have also ran n = 20, which previously was had an error of
~1e-17 near 0 (the max error for n = 20 should be 2.8e-20). It now, based on
the plot, is not losing any precision and has the correct equioscillating
error.