Swap use of alpha and theta in gensolve/py_solve
The typical usage is solve(A - theta*I, alpha*n0), where the thetas are the
roots of the denominator and the alphas are the residues (the variable names
are from the Pusa papers). Technically the theta in diag_add is the negative
of the root, but this should not be too confusing.