Always create plots with LaTeX fonts
This works around some LaTeX issue on my system where the pgf backend can't
find the Bitstream Vera fonts. But we want it to use Computer Modern for the
paper anyway (at least for now), so we might as well do this.

This also fixes the centering of the titles in the nofission plots.